pyproject.toml
[project]
name = "ksdatacloud"
version = "0.1.0"
description = "Data collection and analysis tool for KS Data Cloud solar/battery monitoring platform"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "playwright>=1.52.0",
    "requests>=2.32.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "mypy>=1.8.0",
    "pytest>=8.0.0",
]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
